add method process get payment date processing by personal login

+++ nts.uk/ctx/pr/core/nts.uk.ctx.pr.core.infra/src/main/java/nts/uk/ctx/pr/core/infra/entity/paymentdata/QpdmtPaydayProcessing.java
@@ -23,7 +23,7 @@ public class QpdmtPaydayProcessing extends TableEntity {
 	public int dispAtr;
 	
 	@Column(name = "BONUS_ATR")
-    public short bonusAtr;
+    public int bonusAtr;
 
 	@Column(name = "PROCESSING_NAME")
 	public String processingName;
